Add file of cobot midmap and operation swimlane.
The files are created in the Draw.io enviornment. The first one is a mind map of all cobots involved in the operation, it is created basing on their educational roles and purposes. The second is a swimlane for the operation process. It is similar to both a flow chart and a sequence diagram. It is focusing on the task of different cobot, makes "who does what" more clear. But very different from a state diagram.